refactor(data-table): extract header and footer to their own files
