+++ packages/components/data-table/src/data-table.js
@@ -4,7 +4,8 @@ import isEqual from 'lodash/isEqual';
 import { filterDataAttributes } from '@commercetools-uikit/utils';
 import { usePrevious } from '@commercetools-uikit/hooks';
 import { TableGrid, Header, Body, Row, Footer } from './data-table.styles';
-import { HeaderCell, FooterCell } from './cell';
+import FooterCell from './footer';
+import HeaderCell from './header-cell';
 import DataRow from './data-row';
 import useManualColumnResizing from './use-manual-column-resizing-reducer';
 import ColumnResizingContext from './column-resizing-context';
